Move DefaultOrderedDict to a more natural location.
